FRONT: Reorganized project for allowing more views
