+++ front/src/app/mainVIew/components/inbox/index.js
