+++ front/src/app/mainVIew/components/message/index.js
