+++ front/src/app/mainVIew/components/sidebar/index.js
