+++ front/src/app/mainVIew/index.module.css
@@ -13,5 +13,4 @@
   @apply w-full;
   @apply flex;
   height: 100%;
-
 }
