+++ front/src/index.js
@@ -5,7 +5,7 @@ import { createStore } from "redux";
 import { library } from "@fortawesome/fontawesome-svg-core";
 import { fas } from "@fortawesome/free-solid-svg-icons";
 import { far } from "@fortawesome/free-regular-svg-icons";
-import App from "./components/app";
+import App from "./app/mainVIew";
 import reducer from "./redux/reducer";
 
 import "./stylesheets/basic.css";
